Restructuring code and addition of Phenotyping component
* Restructuring code to separate production and phenotyping.
* Initial structure for Phenotyping component.
* Added plan type to plan details.
