+++ src/app/feature-modules/attempts/components/production/crispr/crispr-attempt-details/crispr-attempt-details.component.html
