+++ src/app/feature-modules/attempts/components/production/crispr/crispr-attempt-details/crispr-attempt-details.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input, EventEmitter, Output } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { CrisprAttempt } from '../../..';
+import { CrisprAttempt } from '../../../..';
 
 @Component({
   selector: 'app-crispr-attempt-details',
