+++ src/app/feature-modules/attempts/components/production/crispr/crispr-attempt/crispr-attempt.component.spec.ts
