+++ src/app/feature-modules/attempts/components/production/crispr/genotype-primers/genotype-primers.component.spec.ts
