+++ src/app/feature-modules/attempts/components/production/crispr/guides/guides.component.css
