+++ src/app/feature-modules/attempts/model/production/crispr/assay.ts
