+++ src/app/feature-modules/attempts/model/production/crispr/crispr-attempt.spec.ts
