+++ src/app/feature-modules/attempts/model/production/crispr/genotype-primer.ts
