+++ src/app/feature-modules/attempts/model/production/crispr/guide.spec.ts
