+++ src/app/feature-modules/attempts/model/production/crispr/reagent.ts
