+++ src/app/feature-modules/attempts/model/production/crispr/strain-injected.ts
